requirements.txt
appdirs==1.4.4
attrs==20.3.0
aws-cdk.alexa-ask==1.92.0
aws-cdk.assets==1.92.0
aws-cdk.aws-accessanalyzer==1.92.0
aws-cdk.aws-acmpca==1.92.0
aws-cdk.aws-amazonmq==1.92.0
aws-cdk.aws-amplify==1.92.0
aws-cdk.aws-apigateway==1.92.0
aws-cdk.aws-apigatewayv2==1.92.0
aws-cdk.aws-appconfig==1.92.0
aws-cdk.aws-appflow==1.92.0
aws-cdk.aws-applicationautoscaling==1.92.0
aws-cdk.aws-applicationinsights==1.92.0
aws-cdk.aws-appmesh==1.92.0
aws-cdk.aws-appstream==1.92.0
aws-cdk.aws-appsync==1.92.0
aws-cdk.aws-athena==1.92.0
aws-cdk.aws-auditmanager==1.92.0
aws-cdk.aws-autoscaling==1.92.0
aws-cdk.aws-autoscaling-common==1.92.0
aws-cdk.aws-autoscaling-hooktargets==1.92.0
aws-cdk.aws-autoscalingplans==1.92.0
aws-cdk.aws-backup==1.92.0
aws-cdk.aws-batch==1.92.0
aws-cdk.aws-budgets==1.92.0
aws-cdk.aws-cassandra==1.92.0
aws-cdk.aws-ce==1.92.0
aws-cdk.aws-certificatemanager==1.92.0
aws-cdk.aws-chatbot==1.92.0
aws-cdk.aws-cloud9==1.92.0
aws-cdk.aws-cloudformation==1.92.0
aws-cdk.aws-cloudfront==1.92.0
aws-cdk.aws-cloudtrail==1.92.0
aws-cdk.aws-cloudwatch==1.92.0
aws-cdk.aws-codeartifact==1.92.0
aws-cdk.aws-codebuild==1.92.0
aws-cdk.aws-codecommit==1.92.0
aws-cdk.aws-codedeploy==1.92.0
aws-cdk.aws-codeguruprofiler==1.92.0
aws-cdk.aws-codegurureviewer==1.92.0
aws-cdk.aws-codepipeline==1.92.0
aws-cdk.aws-codestar==1.92.0
aws-cdk.aws-codestarconnections==1.92.0
aws-cdk.aws-codestarnotifications==1.92.0
aws-cdk.aws-cognito==1.92.0
aws-cdk.aws-config==1.92.0
aws-cdk.aws-databrew==1.92.0
aws-cdk.aws-datapipeline==1.92.0
aws-cdk.aws-datasync==1.92.0
aws-cdk.aws-dax==1.92.0
aws-cdk.aws-detective==1.92.0
aws-cdk.aws-devopsguru==1.92.0
aws-cdk.aws-directoryservice==1.92.0
aws-cdk.aws-dlm==1.92.0
aws-cdk.aws-dms==1.92.0
aws-cdk.aws-docdb==1.92.0
aws-cdk.aws-dynamodb==1.92.0
aws-cdk.aws-ec2==1.92.0
aws-cdk.aws-ecr==1.92.0
aws-cdk.aws-ecr-assets==1.92.0
aws-cdk.aws-ecs==1.92.0
aws-cdk.aws-efs==1.92.0
aws-cdk.aws-eks==1.92.0
aws-cdk.aws-elasticache==1.92.0
aws-cdk.aws-elasticbeanstalk==1.92.0
aws-cdk.aws-elasticloadbalancing==1.92.0
aws-cdk.aws-elasticloadbalancingv2==1.92.0
aws-cdk.aws-elasticsearch==1.92.0
aws-cdk.aws-emr==1.92.0
aws-cdk.aws-emrcontainers==1.92.0
aws-cdk.aws-events==1.92.0
aws-cdk.aws-eventschemas==1.92.0
aws-cdk.aws-fms==1.92.0
aws-cdk.aws-fsx==1.92.0
aws-cdk.aws-gamelift==1.92.0
aws-cdk.aws-globalaccelerator==1.92.0
aws-cdk.aws-glue==1.92.0
aws-cdk.aws-greengrass==1.92.0
aws-cdk.aws-greengrassv2==1.92.0
aws-cdk.aws-guardduty==1.92.0
aws-cdk.aws-iam==1.92.0
aws-cdk.aws-imagebuilder==1.92.0
aws-cdk.aws-inspector==1.92.0
aws-cdk.aws-iot==1.92.0
aws-cdk.aws-iot1click==1.92.0
aws-cdk.aws-iotanalytics==1.92.0
aws-cdk.aws-iotevents==1.92.0
aws-cdk.aws-iotsitewise==1.92.0
aws-cdk.aws-iotthingsgraph==1.92.0
aws-cdk.aws-iotwireless==1.92.0
aws-cdk.aws-ivs==1.92.0
aws-cdk.aws-kendra==1.92.0
aws-cdk.aws-kinesis==1.92.0
aws-cdk.aws-kinesisanalytics==1.92.0
aws-cdk.aws-kinesisfirehose==1.92.0
aws-cdk.aws-kms==1.92.0
aws-cdk.aws-lakeformation==1.92.0
aws-cdk.aws-lambda==1.92.0
aws-cdk.aws-licensemanager==1.92.0
aws-cdk.aws-logs==1.92.0
aws-cdk.aws-lookoutvision==1.92.0
aws-cdk.aws-macie==1.92.0
aws-cdk.aws-managedblockchain==1.92.0
aws-cdk.aws-mediaconnect==1.92.0
aws-cdk.aws-mediaconvert==1.92.0
aws-cdk.aws-medialive==1.92.0
aws-cdk.aws-mediapackage==1.92.0
aws-cdk.aws-mediastore==1.92.0
aws-cdk.aws-msk==1.92.0
aws-cdk.aws-mwaa==1.92.0
aws-cdk.aws-neptune==1.92.0
aws-cdk.aws-networkfirewall==1.92.0
aws-cdk.aws-networkmanager==1.92.0
aws-cdk.aws-opsworks==1.92.0
aws-cdk.aws-opsworkscm==1.92.0
aws-cdk.aws-pinpoint==1.92.0
aws-cdk.aws-pinpointemail==1.92.0
aws-cdk.aws-qldb==1.92.0
aws-cdk.aws-quicksight==1.92.0
aws-cdk.aws-ram==1.92.0
aws-cdk.aws-rds==1.92.0
aws-cdk.aws-redshift==1.92.0
aws-cdk.aws-resourcegroups==1.92.0
aws-cdk.aws-robomaker==1.92.0
aws-cdk.aws-route53==1.92.0
aws-cdk.aws-route53-targets==1.92.0
aws-cdk.aws-route53resolver==1.92.0
aws-cdk.aws-s3==1.92.0
aws-cdk.aws-s3-assets==1.92.0
aws-cdk.aws-sagemaker==1.92.0
aws-cdk.aws-sam==1.92.0
aws-cdk.aws-sdb==1.92.0
aws-cdk.aws-secretsmanager==1.92.0
aws-cdk.aws-securityhub==1.92.0
aws-cdk.aws-servicecatalog==1.92.0
aws-cdk.aws-servicecatalogappregistry==1.92.0
aws-cdk.aws-servicediscovery==1.92.0
aws-cdk.aws-ses==1.92.0
aws-cdk.aws-signer==1.92.0
aws-cdk.aws-sns==1.92.0
aws-cdk.aws-sns-subscriptions==1.92.0
aws-cdk.aws-sqs==1.92.0
aws-cdk.aws-ssm==1.92.0
aws-cdk.aws-sso==1.92.0
aws-cdk.aws-stepfunctions==1.92.0
aws-cdk.aws-synthetics==1.92.0
aws-cdk.aws-timestream==1.92.0
aws-cdk.aws-transfer==1.92.0
aws-cdk.aws-waf==1.92.0
aws-cdk.aws-wafregional==1.92.0
aws-cdk.aws-wafv2==1.92.0
aws-cdk.aws-workspaces==1.92.0
aws-cdk.cloud-assembly-schema==1.92.0
aws-cdk.cloudformation-include==1.92.0
aws-cdk.core==1.92.0
aws-cdk.custom-resources==1.92.0
aws-cdk.cx-api==1.92.0
aws-cdk.lambda-layer-awscli==1.92.0
aws-cdk.lambda-layer-kubectl==1.92.0
aws-cdk.region-info==1.92.0
aws-cdk.yaml-cfn==1.92.0
black==20.8b1
blessed==1.17.6
boto3==1.17.21
botocore==1.20.21
cattrs==1.3.0
chalice==1.22.1
chalice-cdk==0.1.0
click==7.1.2
constructs==3.3.59
enum-compat==0.0.3
inquirer==2.7.0
jmespath==0.10.0
jsii==1.24.0
mypy-extensions==0.4.3
pathspec==0.8.1
publication==0.0.3
pydantic==1.8.1
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.4.1
readchar==2.0.1
regex==2020.11.13
s3transfer==0.3.4
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
